+++ CHANGES.md
@@ -3,6 +3,7 @@
  * Projects:
    * Multi-root Projects! Projects can have multiple subprojects (called Roots) now.
    * Project configuration is now stored inside the project folder as `.dpeproj`.
+   * Directories can be excluded on a per-project basis.
  * Indentation & Whitespace:
    * Added indentation settings for documents.
    * Added whitespace visibility setting.